Modified to allow usage by sub-projects
The modified contributing guidelines including contact and process
information as well as legal requirement for PR's.
It is intended to be linked from other sub-project repositories
to minimize duplication.
Each repository can then include only relevant and specific nuances,
such as a link to `issues`, in its README or CONTRIBUTING.